Parse the IP packet contained in some of the ICMP messages

+++ ui/main_screen.c
@@ -1276,7 +1276,7 @@ void add_elements(main_screen *ms, struct packet *p)
             pinfo->add_pdu(ms->lvw, header, pdata);
         }
     }
-    if (p->perr != NO_ERR && p->len - idx > 0) {
+    if (p->perr != NO_ERR && (int) p->len - idx > 0) {
         header = LV_ADD_HEADER(ms->lvw, "Data", selected[i+1], i + 1);
         add_hexdump(ms->lvw, header, hexmode, p->buf + idx, p->len - idx);
     }
